Fix admin menu in Mybatis
