+++ src/main/java/com/solvd/university/persistence/StudentRepository.java
@@ -5,11 +5,11 @@ import com.solvd.university.domain.Student;
 import java.util.List;
 
 public interface StudentRepository {
+    List<Student> findAll();
+
     void create(Student student);
 
     Student findById(Student student);
-
-    List<Student> findAll();
 
     void update(Student student);
