+++ src/main/java/com/solvd/university/service/AdminService.java
@@ -2,6 +2,4 @@ package com.solvd.university.service;
 
 public interface AdminService {
     void authorizeAdmin();
-
-    void printFullStudentInfo();
 }
