+++ src/main/java/com/solvd/university/service/StudentService.java
@@ -3,6 +3,8 @@ package com.solvd.university.service;
 import com.solvd.university.domain.Student;
 
 public interface StudentService {
+    void printFullStudentInfo();
+
     void enrollStudent();
 
     Student findStudent();
