+++ src/main/java/com/solvd/university/service/impl/commonactions/StudentServiceCommonActions.java
@@ -3,12 +3,27 @@ package com.solvd.university.service.impl.commonactions;
 import com.solvd.university.domain.Student;
 
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Scanner;
 
 import static com.solvd.university.util.ConsoleColors.*;
 import static com.solvd.university.util.MyLogger.MY_LOGGER;
 
 public class StudentServiceCommonActions {
+    public void printWholeStudentInfo(List<Student> studentList) {
+        MY_LOGGER.info(ANSI_GREEN + "Id" + " | " + "Имя" + " | " + "Фамилия" + " | " +
+                "Дата рождения" + " | " + "Средний балл" + ANSI_RESET);
+        for (Student student : studentList) {
+            MY_LOGGER.info(student.getStudentId() + " | " + ANSI_YELLOW +
+                    student.getFirstName() + " | " +
+                    student.getLastName() + " | " +
+                    student.getDateOfBirth() + " | " +
+                    student.getAverageScore() + ANSI_RESET
+            );
+        }
+        System.out.println();
+    }
+
     protected Student addStudent() {
         Scanner scanner = new Scanner(System.in);
         Student student = new Student();
